Some validation to the set before allowing the user to edit the gap time.
Basically the requirement for the algorithm to work is to have lenght times in the set.

original->times.size() == original->lens

otherwise it is impossible to tell what is gap time.

For the current implementation an other requirement is forced: there must be at least 1 length.
Currently we are not recomputing speed and efficiency from the individual times,
just updating the existing values for a different number of lenghts,
but this is impossible if the original number of lengths was 0.

Andrea