Ensure that when a set is edited, the workout totals are synchronised: lenghts, duration and rest (the rest is left untouched if the sets do not include it)
The change in summaryimpl.cpp is to fix a crash in the program, due to the fact
that the length grid had a size bigger than the actual number of lengths.

If the workouts are synchronised, this is not necessary any longer,
but it is left to ensure bad data does not cause a crash.

Andrea