Avoid this QT error:
QObject: Cannot create children for a parent that is in a different thread.
(Parent is QSerialPort(0xf7b4d0), parent's thread is QThread(0x9e18a0), current thread is PodLive(0xf88f40)

This is caused by the fact that the QSerialPort is created (or opened) in a different thread
than the one used for communication.
Basically init() and run() are in different threads; we can simply run init() at the begin of run().

Signed-off-by: Andrea Odetti <mariofutire@gmail.com>