Factor common code to compute the fastest subset in each set.
Change the way partial lengths are handled.
It is easier to read if the duration displayed is adjusted for the target distance (even if this needs an adjustment).

It is now possible to sort by "duration" which avoid the loss of precision in the speed.

Add possibility to zoom (horizontal rubber band).

Signed-off-by: Andrea Odetti <mariofutire@gmail.com>